requirements.txt
alembic==1.10.3
blinker==1.6.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Deprecated==1.2.13
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
Mako==1.2.4
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python-jwt==4.0.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.15.0